extract regression tests
