+++ executor/wasm/Cargo.toml
@@ -27,10 +27,15 @@ thiserror = "2.0"
 tracing = "0.1.40"
 base16 = "0.2.1"
 
+tempfile = { version = "3.10.1", optional = true }
+once_cell = { version = "1.19.0", optional = true }
+fs_extra = { version = "1.3.0", optional = true }
+serde_json = { version = "1.0.127", optional = true }
+itertools = { version = "0.14.0", optional = true }
+
 [dev-dependencies]
-tempfile = "3.10.1"
-once_cell = "1.19.0"
-fs_extra = "1.3.0"
-serde_json = "1.0.127"
-itertools = "0.14.0"
+casper-executor-wasm ={ path = "../wasm", features = ["testing"] }
 casper-wasm = "1.0.0"
+
+[features]
+testing = ["tempfile", "once_cell", "fs_extra", "serde_json", "itertools"]
