+++ executor/wasm/src/testing.rs
@@ -1,0 +1,323 @@
+use std::{
+    env, fs,
+    path::{Path, PathBuf},
+    sync::Arc,
+};
+
+use bytes::Bytes;
+use casper_execution_engine::engine_state::ExecutionEngineV1;
+use casper_executor_wasm_interface::executor::{
+    ExecuteRequest, ExecuteRequestBuilder, ExecuteWithProviderError, ExecuteWithProviderResult,
+};
+use casper_storage::{
+    data_access_layer::{GenesisRequest, GenesisResult},
+    global_state::{
+        self,
+        state::{lmdb::LmdbGlobalState, CommitProvider},
+    },
+    system::runtime_native::Id,
+    AddressGenerator, RuntimeNativeConfig,
+};
+use casper_types::{
+    account::AccountHash, BlockHash, Chainspec, ChainspecRegistry, Digest, GenesisAccount,
+    GenesisConfig, HostFunctionCostsV2, HostFunctionV2, Key, MessageLimits, Motes, Phase,
+    ProtocolVersion, PublicKey, SecretKey, StorageCosts, SystemConfig, Timestamp, TransactionHash,
+    TransactionV1Hash, WasmConfig, WasmV2Config, DEFAULT_WASM_MAX_MEMORY, U512,
+};
+
+use once_cell::sync::Lazy;
+use parking_lot::RwLock;
+use tempfile::TempDir;
+
+use crate::{
+    install::{
+        InstallContractError, InstallContractRequest, InstallContractRequestBuilder,
+        InstallContractResult,
+    },
+    ExecutorConfigBuilder, ExecutorKind, ExecutorV2,
+};
+
+pub static DEFAULT_ACCOUNT_SECRET_KEY: Lazy<SecretKey> =
+    Lazy::new(|| SecretKey::ed25519_from_bytes([199; SecretKey::ED25519_LENGTH]).unwrap());
+pub static DEFAULT_ACCOUNT_PUBLIC_KEY: Lazy<casper_types::PublicKey> =
+    Lazy::new(|| PublicKey::from(&*DEFAULT_ACCOUNT_SECRET_KEY));
+pub static DEFAULT_ACCOUNT_HASH: Lazy<AccountHash> =
+    Lazy::new(|| DEFAULT_ACCOUNT_PUBLIC_KEY.to_account_hash());
+
+pub const CSPR: u64 = 10u64.pow(9);
+
+pub static RUST_WORKSPACE_PATH: Lazy<PathBuf> = Lazy::new(|| {
+    let path = Path::new(env!("CARGO_MANIFEST_DIR"))
+        .parent()
+        .and_then(Path::parent)
+        .expect("CARGO_MANIFEST_DIR should have parent");
+    assert!(
+        path.exists(),
+        "Workspace path {} does not exists",
+        path.display()
+    );
+    path.to_path_buf()
+});
+
+pub static RUST_WORKSPACE_WASM_PATH: Lazy<PathBuf> = Lazy::new(|| {
+    let path = RUST_WORKSPACE_PATH
+        .join("target")
+        .join("wasm32-unknown-unknown")
+        .join("release");
+    assert!(
+        path.exists() || RUST_TOOL_WASM_PATH.exists(),
+        "Rust Wasm path {} does not exists",
+        path.display()
+    );
+    path
+});
+
+// The location of compiled Wasm files if running from within the 'tests' crate generated by the
+// cargo_casper tool, i.e. 'wasm/'.
+pub static RUST_TOOL_WASM_PATH: Lazy<PathBuf> = Lazy::new(|| {
+    env::current_dir()
+        .expect("should get current working dir")
+        .join("wasm")
+});
+
+#[track_caller]
+pub fn read_wasm<P: AsRef<Path>>(filename: P) -> Bytes {
+    let paths = vec![
+        RUST_WORKSPACE_WASM_PATH.clone(),
+        RUST_TOOL_WASM_PATH.clone(),
+    ];
+
+    for path in &paths {
+        let wasm_path = path.join(&filename);
+        match fs::read(wasm_path) {
+            Ok(bytes) => return Bytes::from(bytes),
+            Err(err) => {
+                if err.kind() == std::io::ErrorKind::NotFound {
+                    continue;
+                } else {
+                    panic!(
+                        "Failed to read Wasm file at {}: {}",
+                        filename.as_ref().display(),
+                        err
+                    );
+                }
+            }
+        }
+    }
+
+    panic!(
+        "Failed to find Wasm file at {} in any of the paths: {:?}",
+        filename.as_ref().display(),
+        paths
+    );
+}
+
+pub const TRANSACTION_HASH_BYTES: [u8; 32] = [55; 32];
+pub const TRANSACTION_HASH: TransactionHash =
+    TransactionHash::V1(TransactionV1Hash::from_raw(TRANSACTION_HASH_BYTES));
+pub const DEFAULT_GAS_LIMIT: u64 = 1_000_000 * CSPR;
+pub const DEFAULT_CHAIN_NAME: &str = "casper-test";
+
+// TODO: This is a temporary value, it should be set in the config. Default value from V1 engine
+// does not apply to V2 engine due to different cost structure. Rather than hardcoding it here, we
+// should probably reflect gas costs in a dynamic costs in host function charge. Proper value is
+// pending calculation.
+pub const DEFAULT_GAS_PER_BYTE_COST: u32 = 1_117_587;
+
+pub fn make_address_generator() -> Arc<RwLock<AddressGenerator>> {
+    let id = Id::Transaction(TRANSACTION_HASH);
+    Arc::new(RwLock::new(AddressGenerator::new(
+        &id.seed(),
+        Phase::Session,
+    )))
+}
+
+pub fn base_execute_builder() -> ExecuteRequestBuilder {
+    let chainspec = Chainspec::default();
+    let runtime_native_config = RuntimeNativeConfig::from_chainspec(&chainspec);
+
+    ExecuteRequestBuilder::default()
+        .with_initiator(*DEFAULT_ACCOUNT_HASH)
+        .with_caller_key(Key::Account(*DEFAULT_ACCOUNT_HASH))
+        .with_gas_limit(DEFAULT_GAS_LIMIT)
+        .with_transferred_value(1000)
+        .with_transaction_hash(TRANSACTION_HASH)
+        .with_chain_name(DEFAULT_CHAIN_NAME)
+        .with_block_time(Timestamp::now().into())
+        .with_state_hash(Digest::hash(b"state"))
+        .with_block_height(1)
+        .with_parent_block_hash(BlockHash::new(Digest::hash(b"block1")))
+        .with_runtime_native_config(runtime_native_config)
+}
+
+pub fn base_install_request_builder() -> InstallContractRequestBuilder {
+    let chainspec = Chainspec::default();
+    let runtime_native_config = RuntimeNativeConfig::from_chainspec(&chainspec);
+
+    InstallContractRequestBuilder::default()
+        .with_initiator(*DEFAULT_ACCOUNT_HASH)
+        .with_gas_limit(DEFAULT_GAS_LIMIT)
+        .with_transaction_hash(TRANSACTION_HASH)
+        .with_chain_name(DEFAULT_CHAIN_NAME)
+        .with_block_time(Timestamp::now().into())
+        .with_state_hash(Digest::hash(b"state"))
+        .with_block_height(1)
+        .with_parent_block_hash(BlockHash::new(Digest::hash(b"block1")))
+        .with_runtime_native_config(runtime_native_config)
+}
+
+pub fn make_executor() -> ExecutorV2 {
+    let storage_costs = StorageCosts::new(DEFAULT_GAS_PER_BYTE_COST);
+    let execution_engine_v1 = ExecutionEngineV1::default();
+    let executor_config = ExecutorConfigBuilder::default()
+        .with_memory_limit(DEFAULT_WASM_MAX_MEMORY)
+        .with_executor_kind(ExecutorKind::Compiled)
+        .with_wasm_config(WasmV2Config::default())
+        .with_storage_costs(storage_costs)
+        .with_message_limits(MessageLimits::default())
+        .build()
+        .expect("Should build");
+    ExecutorV2::new(executor_config, Arc::new(execution_engine_v1))
+}
+
+pub fn make_global_state_with_genesis() -> (LmdbGlobalState, Digest, TempDir) {
+    let default_accounts = vec![GenesisAccount::Account {
+        public_key: DEFAULT_ACCOUNT_PUBLIC_KEY.clone(),
+        balance: Motes::new(U512::from(100 * CSPR)),
+        validator: None,
+    }];
+
+    let (global_state, _state_root_hash, _tempdir) =
+        global_state::state::lmdb::make_temporary_global_state([]);
+
+    let genesis_config = GenesisConfig::new(
+        default_accounts,
+        WasmConfig::default(),
+        SystemConfig::default(),
+        10,
+        10,
+        0,
+        Default::default(),
+        14,
+        Timestamp::now().millis(),
+        casper_types::HoldBalanceHandling::Accrued,
+        0,
+        true,
+        StorageCosts::default(),
+    );
+    let genesis_request: GenesisRequest = GenesisRequest::new(
+        Digest::hash("foo"),
+        ProtocolVersion::V2_0_0,
+        genesis_config,
+        ChainspecRegistry::new_with_genesis(b"", b""),
+    );
+    match global_state.genesis(genesis_request) {
+        GenesisResult::Failure(failure) => panic!("Failed to run genesis: {:?}", failure),
+        GenesisResult::Fatal(fatal) => panic!("Fatal error while running genesis: {}", fatal),
+        GenesisResult::Success {
+            post_state_hash,
+            effects: _,
+        } => (global_state, post_state_hash, _tempdir),
+    }
+}
+
+pub fn run_create_contract(
+    executor: &mut ExecutorV2,
+    global_state: &LmdbGlobalState,
+    pre_state_hash: Digest,
+    install_contract_request: InstallContractRequest,
+) -> InstallContractResult {
+    executor
+        .install_contract(pre_state_hash, global_state, install_contract_request)
+        .expect("Succeed")
+}
+
+pub fn expect_successful_execution(
+    executor: &mut ExecutorV2,
+    global_state: &LmdbGlobalState,
+    pre_state_hash: Digest,
+    execute_request: ExecuteRequest,
+) -> ExecuteWithProviderResult {
+    let result =
+        run_wasm_session(executor, global_state, pre_state_hash, execute_request).expect("Succeed");
+
+    if let Some(host_error) = result.host_error {
+        panic!("Host error: {host_error:?}")
+    }
+
+    result
+}
+
+pub fn run_wasm_session(
+    executor: &mut ExecutorV2,
+    global_state: &LmdbGlobalState,
+    pre_state_hash: Digest,
+    execute_request: ExecuteRequest,
+) -> Result<ExecuteWithProviderResult, ExecuteWithProviderError> {
+    executor.execute_with_provider(pre_state_hash, global_state, execute_request)
+}
+
+pub fn call_dummy_host_fn_by_name(
+    host_function_name: &str,
+    gas_limit: u64,
+) -> Result<InstallContractResult, InstallContractError> {
+    let executor = {
+        let execution_engine_v1 = ExecutionEngineV1::default();
+        let default_wasm_config = WasmV2Config::default();
+        let wasm_config = WasmV2Config::new(
+            default_wasm_config.max_memory(),
+            default_wasm_config.opcode_costs(),
+            HostFunctionCostsV2 {
+                read: HostFunctionV2::fixed(1),
+                write: HostFunctionV2::fixed(1),
+                remove: HostFunctionV2::fixed(1),
+                copy_input: HostFunctionV2::fixed(1),
+                ret: HostFunctionV2::fixed(1),
+                create: HostFunctionV2::fixed(1),
+                transfer: HostFunctionV2::fixed(1),
+                env_balance: HostFunctionV2::fixed(1),
+                upgrade: HostFunctionV2::fixed(1),
+                call: HostFunctionV2::fixed(1),
+                print: HostFunctionV2::fixed(1),
+                emit: HostFunctionV2::fixed(1),
+                env_info: HostFunctionV2::fixed(1),
+            },
+        );
+        let executor_config = ExecutorConfigBuilder::default()
+            .with_memory_limit(DEFAULT_WASM_MAX_MEMORY)
+            .with_executor_kind(ExecutorKind::Compiled)
+            .with_wasm_config(wasm_config)
+            .with_storage_costs(StorageCosts::default())
+            .with_message_limits(MessageLimits::default())
+            .build()
+            .expect("Should build");
+        ExecutorV2::new(executor_config, Arc::new(execution_engine_v1))
+    };
+
+    let (global_state, state_root_hash, _tempdir) = make_global_state_with_genesis();
+
+    let address_generator = make_address_generator();
+
+    let input_data = borsh::to_vec(&(host_function_name.to_owned(),))
+        .map(Bytes::from)
+        .unwrap();
+
+    let create_request = base_install_request_builder()
+        .with_initiator(*DEFAULT_ACCOUNT_HASH)
+        .with_gas_limit(gas_limit)
+        .with_transaction_hash(TRANSACTION_HASH)
+        .with_wasm_bytes(read_wasm("vm2_host.wasm"))
+        .with_shared_address_generator(Arc::clone(&address_generator))
+        .with_transferred_value(0)
+        .with_entry_point("new".to_string())
+        .with_input(input_data)
+        .with_chain_name(DEFAULT_CHAIN_NAME)
+        .with_block_time(Timestamp::now().into())
+        .with_state_hash(Digest::from_raw([0; 32]))
+        .with_block_height(1)
+        .with_parent_block_hash(BlockHash::new(Digest::from_raw([0; 32])))
+        .build()
+        .expect("should build");
+
+    executor.install_contract(state_root_hash, &global_state, create_request)
+}
