+++ executor/wasm/tests/integration.rs
@@ -1,184 +1,35 @@
-use std::{
-    env,
-    fs::{self, File},
-    path::{Path, PathBuf},
-    sync::Arc,
-};
+use std::{fs::File, path::Path, sync::Arc};
 
 use bytes::Bytes;
-use casper_execution_engine::engine_state::ExecutionEngineV1;
-use casper_executor_wasm::{
-    install::{
-        InstallContractError, InstallContractRequest, InstallContractRequestBuilder,
-        InstallContractResult,
-    },
-    ExecutorConfigBuilder, ExecutorKind, ExecutorV2,
-};
+use casper_executor_wasm::install::{InstallContractError, InstallContractRequest};
 use casper_executor_wasm_common::error::CallError;
-use casper_executor_wasm_interface::{
-    executor::{
-        ExecuteError, ExecuteRequest, ExecuteRequestBuilder, ExecuteWithProviderError,
-        ExecuteWithProviderResult, ExecutionKind,
-    },
-    WasmPreparationError,
+use casper_executor_wasm_interface::executor::{
+    ExecuteError, ExecuteWithProviderError, ExecutionKind,
 };
 use casper_storage::{
     data_access_layer::{
         prefixed_values::{PrefixedValuesRequest, PrefixedValuesResult},
-        GenesisRequest, GenesisResult, MessageTopicsRequest, MessageTopicsResult, QueryRequest,
-        QueryResult,
+        MessageTopicsRequest, MessageTopicsResult, QueryRequest, QueryResult,
     },
     global_state::{
-        self,
         state::{lmdb::LmdbGlobalState, CommitProvider, StateProvider},
         transaction_source::lmdb::LmdbEnvironment,
         trie_store::lmdb::LmdbTrieStore,
     },
-    system::runtime_native::Id,
-    AddressGenerator, KeyPrefix, RuntimeNativeConfig,
+    KeyPrefix,
 };
 use casper_types::{
-    account::AccountHash, execution::RetValue, BlockHash, Chainspec, ChainspecRegistry, Digest,
-    EntityAddr, GenesisAccount, GenesisConfig, HostFunctionCostsV2, HostFunctionV2, Key,
-    MessageLimits, Motes, Phase, ProtocolVersion, PublicKey, SecretKey, StorageCosts, StoredValue,
-    SystemConfig, Timestamp, TransactionHash, TransactionV1Hash, WasmConfig, WasmV2Config,
-    DEFAULT_WASM_MAX_MEMORY, U512,
+    execution::RetValue, BlockHash, Digest, EntityAddr, Key, StoredValue, Timestamp,
 };
-use casper_wasm::builder;
 use fs_extra::dir;
 use itertools::Itertools;
-use once_cell::sync::Lazy;
-use parking_lot::RwLock;
-use tempfile::TempDir;
 
-static DEFAULT_ACCOUNT_SECRET_KEY: Lazy<SecretKey> =
-    Lazy::new(|| SecretKey::ed25519_from_bytes([199; SecretKey::ED25519_LENGTH]).unwrap());
-static DEFAULT_ACCOUNT_PUBLIC_KEY: Lazy<casper_types::PublicKey> =
-    Lazy::new(|| PublicKey::from(&*DEFAULT_ACCOUNT_SECRET_KEY));
-static DEFAULT_ACCOUNT_HASH: Lazy<AccountHash> =
-    Lazy::new(|| DEFAULT_ACCOUNT_PUBLIC_KEY.to_account_hash());
-
-const CONTRACT_EE_966_REGRESSION: &str = "vm2_ee_966_regression.wasm";
-const CSPR: u64 = 10u64.pow(9);
-
-static RUST_WORKSPACE_PATH: Lazy<PathBuf> = Lazy::new(|| {
-    let path = Path::new(env!("CARGO_MANIFEST_DIR"))
-        .parent()
-        .and_then(Path::parent)
-        .expect("CARGO_MANIFEST_DIR should have parent");
-    assert!(
-        path.exists(),
-        "Workspace path {} does not exists",
-        path.display()
-    );
-    path.to_path_buf()
-});
-
-static RUST_WORKSPACE_WASM_PATH: Lazy<PathBuf> = Lazy::new(|| {
-    let path = RUST_WORKSPACE_PATH
-        .join("target")
-        .join("wasm32-unknown-unknown")
-        .join("release");
-    assert!(
-        path.exists() || RUST_TOOL_WASM_PATH.exists(),
-        "Rust Wasm path {} does not exists",
-        path.display()
-    );
-    path
-});
-// The location of compiled Wasm files if running from within the 'tests' crate generated by the
-// cargo_casper tool, i.e. 'wasm/'.
-static RUST_TOOL_WASM_PATH: Lazy<PathBuf> = Lazy::new(|| {
-    env::current_dir()
-        .expect("should get current working dir")
-        .join("wasm")
-});
-
-#[track_caller]
-fn read_wasm<P: AsRef<Path>>(filename: P) -> Bytes {
-    let paths = vec![
-        RUST_WORKSPACE_WASM_PATH.clone(),
-        RUST_TOOL_WASM_PATH.clone(),
-    ];
-
-    for path in &paths {
-        let wasm_path = path.join(&filename);
-        match fs::read(wasm_path) {
-            Ok(bytes) => return Bytes::from(bytes),
-            Err(err) => {
-                if err.kind() == std::io::ErrorKind::NotFound {
-                    continue;
-                } else {
-                    panic!(
-                        "Failed to read Wasm file at {}: {}",
-                        filename.as_ref().display(),
-                        err
-                    );
-                }
-            }
-        }
-    }
-
-    panic!(
-        "Failed to find Wasm file at {} in any of the paths: {:?}",
-        filename.as_ref().display(),
-        paths
-    );
-}
-
-const TRANSACTION_HASH_BYTES: [u8; 32] = [55; 32];
-const TRANSACTION_HASH: TransactionHash =
-    TransactionHash::V1(TransactionV1Hash::from_raw(TRANSACTION_HASH_BYTES));
-const DEFAULT_GAS_LIMIT: u64 = 1_000_000 * CSPR;
-const DEFAULT_CHAIN_NAME: &str = "casper-test";
-
-// TODO: This is a temporary value, it should be set in the config. Default value from V1 engine
-// does not apply to V2 engine due to different cost structure. Rather than hardcoding it here, we
-// should probably reflect gas costs in a dynamic costs in host function charge. Proper value is
-// pending calculation.
-const DEFAULT_GAS_PER_BYTE_COST: u32 = 1_117_587;
-
-fn make_address_generator() -> Arc<RwLock<AddressGenerator>> {
-    let id = Id::Transaction(TRANSACTION_HASH);
-    Arc::new(RwLock::new(AddressGenerator::new(
-        &id.seed(),
-        Phase::Session,
-    )))
-}
-
-fn base_execute_builder() -> ExecuteRequestBuilder {
-    let chainspec = Chainspec::default();
-    let runtime_native_config = RuntimeNativeConfig::from_chainspec(&chainspec);
-
-    ExecuteRequestBuilder::default()
-        .with_initiator(*DEFAULT_ACCOUNT_HASH)
-        .with_caller_key(Key::Account(*DEFAULT_ACCOUNT_HASH))
-        .with_gas_limit(DEFAULT_GAS_LIMIT)
-        .with_transferred_value(1000)
-        .with_transaction_hash(TRANSACTION_HASH)
-        .with_chain_name(DEFAULT_CHAIN_NAME)
-        .with_block_time(Timestamp::now().into())
-        .with_state_hash(Digest::hash(b"state"))
-        .with_block_height(1)
-        .with_parent_block_hash(BlockHash::new(Digest::hash(b"block1")))
-        .with_runtime_native_config(runtime_native_config)
-}
-
-fn base_install_request_builder() -> InstallContractRequestBuilder {
-    let chainspec = Chainspec::default();
-    let runtime_native_config = RuntimeNativeConfig::from_chainspec(&chainspec);
-
-    InstallContractRequestBuilder::default()
-        .with_initiator(*DEFAULT_ACCOUNT_HASH)
-        .with_gas_limit(DEFAULT_GAS_LIMIT)
-        .with_transaction_hash(TRANSACTION_HASH)
-        .with_chain_name(DEFAULT_CHAIN_NAME)
-        .with_block_time(Timestamp::now().into())
-        .with_state_hash(Digest::hash(b"state"))
-        .with_block_height(1)
-        .with_parent_block_hash(BlockHash::new(Digest::hash(b"block1")))
-        .with_runtime_native_config(runtime_native_config)
-}
+use casper_executor_wasm::testing::{
+    base_execute_builder, base_install_request_builder, call_dummy_host_fn_by_name,
+    expect_successful_execution, make_address_generator, make_executor,
+    make_global_state_with_genesis, read_wasm, run_create_contract, DEFAULT_ACCOUNT_HASH,
+    DEFAULT_CHAIN_NAME, DEFAULT_GAS_LIMIT, TRANSACTION_HASH,
+};
 
 #[test]
 fn harness() {
@@ -241,20 +92,6 @@ fn harness() {
         state_root_hash,
         execute_request,
     );
-}
-
-pub(crate) fn make_executor() -> ExecutorV2 {
-    let storage_costs = StorageCosts::new(DEFAULT_GAS_PER_BYTE_COST);
-    let execution_engine_v1 = ExecutionEngineV1::default();
-    let executor_config = ExecutorConfigBuilder::default()
-        .with_memory_limit(DEFAULT_WASM_MAX_MEMORY)
-        .with_executor_kind(ExecutorKind::Compiled)
-        .with_wasm_config(WasmV2Config::default())
-        .with_storage_costs(storage_costs)
-        .with_message_limits(MessageLimits::default())
-        .build()
-        .expect("Should build");
-    ExecutorV2::new(executor_config, Arc::new(execution_engine_v1))
 }
 
 #[test]
@@ -418,47 +255,6 @@ fn cep18() {
     assert_eq!(messages[1].topic_name(), "Transfer");
     assert_eq!(messages[1].topic_index(), 1);
     assert_eq!(messages[1].block_index(), 1);
-}
-
-fn make_global_state_with_genesis() -> (LmdbGlobalState, Digest, TempDir) {
-    let default_accounts = vec![GenesisAccount::Account {
-        public_key: DEFAULT_ACCOUNT_PUBLIC_KEY.clone(),
-        balance: Motes::new(U512::from(100 * CSPR)),
-        validator: None,
-    }];
-
-    let (global_state, _state_root_hash, _tempdir) =
-        global_state::state::lmdb::make_temporary_global_state([]);
-
-    let genesis_config = GenesisConfig::new(
-        default_accounts,
-        WasmConfig::default(),
-        SystemConfig::default(),
-        10,
-        10,
-        0,
-        Default::default(),
-        14,
-        Timestamp::now().millis(),
-        casper_types::HoldBalanceHandling::Accrued,
-        0,
-        true,
-        StorageCosts::default(),
-    );
-    let genesis_request: GenesisRequest = GenesisRequest::new(
-        Digest::hash("foo"),
-        ProtocolVersion::V2_0_0,
-        genesis_config,
-        ChainspecRegistry::new_with_genesis(b"", b""),
-    );
-    match global_state.genesis(genesis_request) {
-        GenesisResult::Failure(failure) => panic!("Failed to run genesis: {:?}", failure),
-        GenesisResult::Fatal(fatal) => panic!("Fatal error while running genesis: {}", fatal),
-        GenesisResult::Success {
-            post_state_hash,
-            effects: _,
-        } => (global_state, post_state_hash, _tempdir),
-    }
 }
 
 #[test]
@@ -641,42 +437,6 @@ fn upgradable() {
     let _ = state_root_hash;
 }
 
-fn run_create_contract(
-    executor: &mut ExecutorV2,
-    global_state: &LmdbGlobalState,
-    pre_state_hash: Digest,
-    install_contract_request: InstallContractRequest,
-) -> InstallContractResult {
-    executor
-        .install_contract(pre_state_hash, global_state, install_contract_request)
-        .expect("Succeed")
-}
-
-fn expect_successful_execution(
-    executor: &mut ExecutorV2,
-    global_state: &LmdbGlobalState,
-    pre_state_hash: Digest,
-    execute_request: ExecuteRequest,
-) -> ExecuteWithProviderResult {
-    let result =
-        run_wasm_session(executor, global_state, pre_state_hash, execute_request).expect("Succeed");
-
-    if let Some(host_error) = result.host_error {
-        panic!("Host error: {host_error:?}")
-    }
-
-    result
-}
-
-fn run_wasm_session(
-    executor: &mut ExecutorV2,
-    global_state: &LmdbGlobalState,
-    pre_state_hash: Digest,
-    execute_request: ExecuteRequest,
-) -> Result<ExecuteWithProviderResult, ExecuteWithProviderError> {
-    executor.execute_with_provider(pre_state_hash, global_state, execute_request)
-}
-
 #[test]
 fn backwards_compatibility() {
     let (global_state, post_state_hash, _temp) = {
@@ -834,85 +594,18 @@ fn backwards_compatibility() {
     expect_successful_execution(&mut executor, &global_state, state_root_hash, call_request);
 }
 
-// host function tests
-
-fn call_dummy_host_fn_by_name(
-    host_function_name: &str,
-    gas_limit: u64,
-) -> Result<InstallContractResult, InstallContractError> {
-    let executor = {
-        let execution_engine_v1 = ExecutionEngineV1::default();
-        let default_wasm_config = WasmV2Config::default();
-        let wasm_config = WasmV2Config::new(
-            default_wasm_config.max_memory(),
-            default_wasm_config.opcode_costs(),
-            HostFunctionCostsV2 {
-                read: HostFunctionV2::fixed(1),
-                write: HostFunctionV2::fixed(1),
-                remove: HostFunctionV2::fixed(1),
-                copy_input: HostFunctionV2::fixed(1),
-                ret: HostFunctionV2::fixed(1),
-                create: HostFunctionV2::fixed(1),
-                transfer: HostFunctionV2::fixed(1),
-                env_balance: HostFunctionV2::fixed(1),
-                upgrade: HostFunctionV2::fixed(1),
-                call: HostFunctionV2::fixed(1),
-                print: HostFunctionV2::fixed(1),
-                emit: HostFunctionV2::fixed(1),
-                env_info: HostFunctionV2::fixed(1),
-            },
-        );
-        let executor_config = ExecutorConfigBuilder::default()
-            .with_memory_limit(DEFAULT_WASM_MAX_MEMORY)
-            .with_executor_kind(ExecutorKind::Compiled)
-            .with_wasm_config(wasm_config)
-            .with_storage_costs(StorageCosts::default())
-            .with_message_limits(MessageLimits::default())
-            .build()
-            .expect("Should build");
-        ExecutorV2::new(executor_config, Arc::new(execution_engine_v1))
-    };
-
-    let (global_state, state_root_hash, _tempdir) = make_global_state_with_genesis();
-
-    let address_generator = make_address_generator();
-
-    let input_data = borsh::to_vec(&(host_function_name.to_owned(),))
-        .map(Bytes::from)
-        .unwrap();
-
-    let create_request = base_install_request_builder()
-        .with_initiator(*DEFAULT_ACCOUNT_HASH)
-        .with_gas_limit(gas_limit)
-        .with_transaction_hash(TRANSACTION_HASH)
-        .with_wasm_bytes(read_wasm("vm2_host.wasm"))
-        .with_shared_address_generator(Arc::clone(&address_generator))
-        .with_transferred_value(0)
-        .with_entry_point("new".to_string())
-        .with_input(input_data)
-        .with_chain_name(DEFAULT_CHAIN_NAME)
-        .with_block_time(Timestamp::now().into())
-        .with_state_hash(Digest::from_raw([0; 32]))
-        .with_block_height(1)
-        .with_parent_block_hash(BlockHash::new(Digest::from_raw([0; 32])))
-        .build()
-        .expect("should build");
-
-    executor.install_contract(state_root_hash, &global_state, create_request)
-}
-
-fn assert_consumes_gas(host_function_name: &str) {
-    let result = call_dummy_host_fn_by_name(host_function_name, 1);
-    assert!(result.is_err_and(|e| matches!(
-        e,
-        InstallContractError::Constructor {
-            host_error: CallError::CalleeGasDepleted,
-        }
-    )));
-}
-
 #[test]
 fn host_functions_consume_gas() {
+    fn assert_consumes_gas(host_function_name: &str) {
+        let result = call_dummy_host_fn_by_name(host_function_name, 1);
+        assert!(result.is_err_and(|e| matches!(
+            e,
+            InstallContractError::Constructor {
+                host_error: CallError::CalleeGasDepleted,
+            }
+        )));
+    }
+
     assert_consumes_gas("get_caller");
     assert_consumes_gas("get_block_time");
     assert_consumes_gas("get_transferred_value");
@@ -927,84 +620,6 @@ fn host_functions_consume_gas() {
     assert_consumes_gas("upgrade");
     assert_consumes_gas("write");
 }
-
-#[allow(dead_code)]
-fn write_n_bytes_at_limit(
-    bytes_len: u64,
-    gas_limit: u64,
-) -> Result<InstallContractResult, InstallContractError> {
-    let executor = {
-        let execution_engine_v1 = ExecutionEngineV1::default();
-        let default_wasm_config = WasmV2Config::default();
-        let wasm_config = WasmV2Config::new(
-            default_wasm_config.max_memory(),
-            default_wasm_config.opcode_costs(),
-            HostFunctionCostsV2 {
-                read: HostFunctionV2::fixed(0),
-                write: HostFunctionV2::fixed(0),
-                remove: HostFunctionV2::fixed(0),
-                copy_input: HostFunctionV2::fixed(0),
-                ret: HostFunctionV2::fixed(0),
-                create: HostFunctionV2::fixed(0),
-                transfer: HostFunctionV2::fixed(0),
-                env_balance: HostFunctionV2::fixed(0),
-                upgrade: HostFunctionV2::fixed(0),
-                call: HostFunctionV2::fixed(0),
-                print: HostFunctionV2::fixed(0),
-                emit: HostFunctionV2::fixed(0),
-                env_info: HostFunctionV2::fixed(0),
-            },
-        );
-        let executor_config = ExecutorConfigBuilder::default()
-            .with_memory_limit(17)
-            .with_executor_kind(ExecutorKind::Compiled)
-            .with_wasm_config(wasm_config)
-            .with_storage_costs(StorageCosts::new(1))
-            .with_message_limits(MessageLimits::default())
-            .build()
-            .expect("Should build");
-        ExecutorV2::new(executor_config, Arc::new(execution_engine_v1))
-    };
-
-    let (global_state, state_root_hash, _tempdir) = make_global_state_with_genesis();
-
-    let address_generator = make_address_generator();
-
-    let input_data = borsh::to_vec(&(bytes_len,)).map(Bytes::from).unwrap();
-
-    let create_request = InstallContractRequestBuilder::default()
-        .with_initiator(*DEFAULT_ACCOUNT_HASH)
-        .with_gas_limit(gas_limit)
-        .with_transaction_hash(TRANSACTION_HASH)
-        .with_wasm_bytes(read_wasm("vm2_host.wasm"))
-        .with_shared_address_generator(Arc::clone(&address_generator))
-        .with_transferred_value(0)
-        .with_entry_point("new_with_write".to_string())
-        .with_input(input_data)
-        .with_chain_name(DEFAULT_CHAIN_NAME)
-        .with_block_time(Timestamp::now().into())
-        .with_state_hash(Digest::from_raw([0; 32]))
-        .with_block_height(1)
-        .with_parent_block_hash(BlockHash::new(Digest::from_raw([0; 32])))
-        .build()
-        .expect("should build");
-
-    executor.install_contract(state_root_hash, &global_state, create_request)
-}
-
-// #[test]
-// fn consume_gas_on_write() {
-//     let successful_write = write_n_bytes_at_limit(50, 10_000);
-//     assert!(successful_write.is_ok());
-
-//     let out_of_gas_write_exceeded_gas_limit = write_n_bytes_at_limit(50, 10);
-//     assert!(out_of_gas_write_exceeded_gas_limit.is_err_and(|e| match e {
-//         InstallContractError::Constructor {
-//             host_error: HostError::CalleeGasDepleted,
-//         } => true,
-//         _ => false,
-//     }));
-// }
 
 #[test]
 fn non_existing_smart_contract_does_not_panic() {
@@ -1120,406 +735,4 @@ fn casper_return_writes_to_execution_journal() {
         &expected_key,
         "Ret transform should be under the contract key"
     );
-}
-
-#[test]
-fn argument_size_exceeds_memory_limit() {
-    use casper_executor_wasm_interface::executor::ExecuteError;
-    let executor = {
-        let storage_costs = StorageCosts::new(DEFAULT_GAS_PER_BYTE_COST);
-        let execution_engine_v1 = ExecutionEngineV1::default();
-        // Config with one page of mem (64KiB)
-        let executor_config = ExecutorConfigBuilder::default()
-            .with_memory_limit(1)
-            .with_executor_kind(ExecutorKind::Compiled)
-            .with_wasm_config(WasmV2Config::default())
-            .with_storage_costs(storage_costs)
-            .with_message_limits(MessageLimits::default())
-            .build()
-            .expect("Should build");
-        ExecutorV2::new(executor_config, Arc::new(execution_engine_v1))
-    };
-    let (global_state, state_root_hash, _tempdir) = make_global_state_with_genesis();
-    let address_generator = make_address_generator();
-    let chainspec = Chainspec::default();
-    let runtime_native_config = RuntimeNativeConfig::from_chainspec(&chainspec);
-    // Create an input larger than 1 page
-    let large_input = Bytes::from(vec![0u8; 70_000]);
-    let execute_request = ExecuteRequestBuilder::default()
-        .with_initiator(*DEFAULT_ACCOUNT_HASH)
-        .with_caller_key(Key::Account(*DEFAULT_ACCOUNT_HASH))
-        .with_gas_limit(DEFAULT_GAS_LIMIT)
-        .with_transferred_value(0)
-        .with_transaction_hash(TRANSACTION_HASH)
-        .with_target(ExecutionKind::SessionBytes(read_wasm("vm2_cep18.wasm")))
-        .with_input(large_input)
-        .with_shared_address_generator(address_generator)
-        .with_chain_name(DEFAULT_CHAIN_NAME)
-        .with_block_time(Timestamp::now().into())
-        .with_state_hash(state_root_hash)
-        .with_block_height(1)
-        .with_parent_block_hash(BlockHash::new(Digest::hash(b"block1")))
-        .with_runtime_native_config(runtime_native_config)
-        .build()
-        .expect("should build");
-    let result = executor.execute_with_provider(state_root_hash, &global_state, execute_request);
-    match result {
-        Err(ExecuteWithProviderError::Execute(ExecuteError::ArgumentSizeExceedsMemory {
-            argument_size,
-            memory_limit,
-        })) => {
-            assert!(argument_size > (memory_limit as usize * 65536));
-        }
-        other => panic!("Expected ArgumentSizeExceedsMemory error, got: {:?}", other),
-    }
-}
-
-fn make_session_code_with_memory_pages(initial_pages: u32, max_pages: Option<u32>) -> Bytes {
-    let module = builder::module()
-        .function()
-        // A signature with 0 params and no return type
-        .signature()
-        .build()
-        .body()
-        .build()
-        .build()
-        // Export above function
-        .export()
-        .field("call")
-        .build()
-        // Memory section is mandatory
-        .memory()
-        // Produces entry `(memory (0) initial_pages [max_pages])`
-        .with_min(initial_pages)
-        .with_max(max_pages)
-        .build()
-        .build();
-    casper_wasm::serialize(module)
-        .expect("should serialize")
-        .into()
-}
-
-#[test]
-fn should_run_ee_966_with_zero_min_and_zero_max_memory() {
-    // A contract that has initial memory pages of 0 and maximum memory pages of 0 is valid
-    let session_code = make_session_code_with_memory_pages(0, Some(0));
-
-    let mut executor = make_executor();
-    let (global_state, state_root_hash, _tempdir) = make_global_state_with_genesis();
-    let address_generator = make_address_generator();
-
-    let runtime_native_config = RuntimeNativeConfig::from_chainspec(&Chainspec::default());
-    let execute_request = ExecuteRequestBuilder::default()
-        .with_initiator(*DEFAULT_ACCOUNT_HASH)
-        .with_caller_key(Key::Account(*DEFAULT_ACCOUNT_HASH))
-        .with_gas_limit(DEFAULT_GAS_LIMIT)
-        .with_transferred_value(0)
-        .with_transaction_hash(TRANSACTION_HASH)
-        .with_target(ExecutionKind::SessionBytes(session_code))
-        .with_input(Bytes::new())
-        .with_shared_address_generator(address_generator)
-        .with_chain_name(DEFAULT_CHAIN_NAME)
-        .with_block_time(Timestamp::now().into())
-        .with_state_hash(state_root_hash)
-        .with_block_height(1)
-        .with_parent_block_hash(BlockHash::new(Digest::hash(b"block1")))
-        .with_runtime_native_config(runtime_native_config)
-        .build()
-        .expect("should build");
-
-    expect_successful_execution(
-        &mut executor,
-        &global_state,
-        state_root_hash,
-        execute_request,
-    );
-}
-
-#[test]
-fn should_run_ee_966_cant_have_too_much_initial_memory() {
-    // Set initial memory to max + 1
-    let session_code = make_session_code_with_memory_pages(DEFAULT_WASM_MAX_MEMORY + 1, None);
-
-    let mut executor = make_executor();
-    let (global_state, state_root_hash, _tempdir) = make_global_state_with_genesis();
-    let address_generator = make_address_generator();
-
-    let runtime_native_config = RuntimeNativeConfig::from_chainspec(&Chainspec::default());
-    let execute_request = ExecuteRequestBuilder::default()
-        .with_initiator(*DEFAULT_ACCOUNT_HASH)
-        .with_caller_key(Key::Account(*DEFAULT_ACCOUNT_HASH))
-        .with_gas_limit(DEFAULT_GAS_LIMIT)
-        .with_transferred_value(0)
-        .with_transaction_hash(TRANSACTION_HASH)
-        .with_target(ExecutionKind::SessionBytes(session_code))
-        .with_input(Bytes::new())
-        .with_shared_address_generator(address_generator)
-        .with_chain_name(DEFAULT_CHAIN_NAME)
-        .with_block_time(Timestamp::now().into())
-        .with_state_hash(state_root_hash)
-        .with_block_height(1)
-        .with_parent_block_hash(BlockHash::new(Digest::hash(b"block1")))
-        .with_runtime_native_config(runtime_native_config)
-        .build()
-        .expect("should build");
-
-    let result = run_wasm_session(
-        &mut executor,
-        &global_state,
-        state_root_hash,
-        execute_request,
-    );
-
-    assert!(matches!(
-        result,
-        Err(ExecuteWithProviderError::Execute(
-            ExecuteError::WasmPreparation(WasmPreparationError::Memory(_))
-        ))
-    ));
-}
-
-#[test]
-fn should_run_ee_966_cant_have_too_much_max_memory() {
-    let session_code = make_session_code_with_memory_pages(0, Some(DEFAULT_WASM_MAX_MEMORY + 1));
-
-    let mut executor = make_executor();
-    let (global_state, state_root_hash, _tempdir) = make_global_state_with_genesis();
-    let address_generator = make_address_generator();
-
-    let runtime_native_config = RuntimeNativeConfig::from_chainspec(&Chainspec::default());
-    let execute_request = ExecuteRequestBuilder::default()
-        .with_initiator(*DEFAULT_ACCOUNT_HASH)
-        .with_caller_key(Key::Account(*DEFAULT_ACCOUNT_HASH))
-        .with_gas_limit(DEFAULT_GAS_LIMIT)
-        .with_transferred_value(0)
-        .with_transaction_hash(TRANSACTION_HASH)
-        .with_target(ExecutionKind::SessionBytes(session_code))
-        .with_input(Bytes::new())
-        .with_shared_address_generator(address_generator)
-        .with_chain_name(DEFAULT_CHAIN_NAME)
-        .with_block_time(Timestamp::now().into())
-        .with_state_hash(state_root_hash)
-        .with_block_height(1)
-        .with_parent_block_hash(BlockHash::new(Digest::hash(b"block1")))
-        .with_runtime_native_config(runtime_native_config)
-        .build()
-        .expect("should build");
-
-    let result = run_wasm_session(
-        &mut executor,
-        &global_state,
-        state_root_hash,
-        execute_request,
-    );
-
-    assert!(matches!(
-        result,
-        Err(ExecuteWithProviderError::Execute(
-            ExecuteError::WasmPreparation(WasmPreparationError::Instantiation(_))
-        ))
-    ));
-}
-
-#[test]
-fn should_run_ee_966_cant_have_way_too_much_max_memory() {
-    let session_code = make_session_code_with_memory_pages(0, Some(DEFAULT_WASM_MAX_MEMORY * 3));
-
-    let mut executor = make_executor();
-    let (global_state, state_root_hash, _tempdir) = make_global_state_with_genesis();
-    let address_generator = make_address_generator();
-
-    let runtime_native_config = RuntimeNativeConfig::from_chainspec(&Chainspec::default());
-    let execute_request = ExecuteRequestBuilder::default()
-        .with_initiator(*DEFAULT_ACCOUNT_HASH)
-        .with_caller_key(Key::Account(*DEFAULT_ACCOUNT_HASH))
-        .with_gas_limit(DEFAULT_GAS_LIMIT)
-        .with_transferred_value(0)
-        .with_transaction_hash(TRANSACTION_HASH)
-        .with_target(ExecutionKind::SessionBytes(session_code))
-        .with_input(Bytes::new())
-        .with_shared_address_generator(address_generator)
-        .with_chain_name(DEFAULT_CHAIN_NAME)
-        .with_block_time(Timestamp::now().into())
-        .with_state_hash(state_root_hash)
-        .with_block_height(1)
-        .with_parent_block_hash(BlockHash::new(Digest::hash(b"block1")))
-        .with_runtime_native_config(runtime_native_config)
-        .build()
-        .expect("should build");
-
-    let result = run_wasm_session(
-        &mut executor,
-        &global_state,
-        state_root_hash,
-        execute_request,
-    );
-
-    assert!(matches!(
-        result,
-        Err(ExecuteWithProviderError::Execute(
-            ExecuteError::WasmPreparation(WasmPreparationError::Instantiation(_))
-        ))
-    ));
-}
-
-#[test]
-fn should_run_ee_966_cant_have_larger_initial_than_max_memory() {
-    let session_code = make_session_code_with_memory_pages(DEFAULT_WASM_MAX_MEMORY, Some(0));
-
-    let mut executor = make_executor();
-    let (global_state, state_root_hash, _tempdir) = make_global_state_with_genesis();
-    let address_generator = make_address_generator();
-
-    let runtime_native_config = RuntimeNativeConfig::from_chainspec(&Chainspec::default());
-    let execute_request = ExecuteRequestBuilder::default()
-        .with_initiator(*DEFAULT_ACCOUNT_HASH)
-        .with_caller_key(Key::Account(*DEFAULT_ACCOUNT_HASH))
-        .with_gas_limit(DEFAULT_GAS_LIMIT)
-        .with_transferred_value(0)
-        .with_transaction_hash(TRANSACTION_HASH)
-        .with_target(ExecutionKind::SessionBytes(session_code))
-        .with_input(Bytes::new())
-        .with_shared_address_generator(address_generator)
-        .with_chain_name(DEFAULT_CHAIN_NAME)
-        .with_block_time(Timestamp::now().into())
-        .with_state_hash(state_root_hash)
-        .with_block_height(1)
-        .with_parent_block_hash(BlockHash::new(Digest::hash(b"block1")))
-        .with_runtime_native_config(runtime_native_config)
-        .build()
-        .expect("should build");
-
-    let result = run_wasm_session(
-        &mut executor,
-        &global_state,
-        state_root_hash,
-        execute_request,
-    );
-
-    assert!(matches!(
-        result,
-        Err(ExecuteWithProviderError::Execute(
-            ExecuteError::WasmPreparation(WasmPreparationError::Compile(_))
-        ))
-    ));
-}
-
-#[test]
-fn should_run_ee_966_should_request_exactly_maximum_as_initial() {
-    let session_code = make_session_code_with_memory_pages(DEFAULT_WASM_MAX_MEMORY, None);
-
-    let mut executor = make_executor();
-    let (global_state, state_root_hash, _tempdir) = make_global_state_with_genesis();
-    let address_generator = make_address_generator();
-
-    let runtime_native_config = RuntimeNativeConfig::from_chainspec(&Chainspec::default());
-    let execute_request = ExecuteRequestBuilder::default()
-        .with_initiator(*DEFAULT_ACCOUNT_HASH)
-        .with_caller_key(Key::Account(*DEFAULT_ACCOUNT_HASH))
-        .with_gas_limit(DEFAULT_GAS_LIMIT)
-        .with_transferred_value(0)
-        .with_transaction_hash(TRANSACTION_HASH)
-        .with_target(ExecutionKind::SessionBytes(session_code))
-        .with_input(Bytes::new())
-        .with_shared_address_generator(address_generator)
-        .with_chain_name(DEFAULT_CHAIN_NAME)
-        .with_block_time(Timestamp::now().into())
-        .with_state_hash(state_root_hash)
-        .with_block_height(1)
-        .with_parent_block_hash(BlockHash::new(Digest::hash(b"block1")))
-        .with_runtime_native_config(runtime_native_config)
-        .build()
-        .expect("should build");
-
-    let result = run_wasm_session(
-        &mut executor,
-        &global_state,
-        state_root_hash,
-        execute_request,
-    );
-
-    assert!(result.is_ok());
-}
-
-#[test]
-fn should_run_ee_966_should_request_exactly_maximum() {
-    let session_code =
-        make_session_code_with_memory_pages(DEFAULT_WASM_MAX_MEMORY, Some(DEFAULT_WASM_MAX_MEMORY));
-
-    let mut executor = make_executor();
-    let (global_state, state_root_hash, _tempdir) = make_global_state_with_genesis();
-    let address_generator = make_address_generator();
-
-    let runtime_native_config = RuntimeNativeConfig::from_chainspec(&Chainspec::default());
-    let execute_request = ExecuteRequestBuilder::default()
-        .with_initiator(*DEFAULT_ACCOUNT_HASH)
-        .with_caller_key(Key::Account(*DEFAULT_ACCOUNT_HASH))
-        .with_gas_limit(DEFAULT_GAS_LIMIT)
-        .with_transferred_value(0)
-        .with_transaction_hash(TRANSACTION_HASH)
-        .with_target(ExecutionKind::SessionBytes(session_code))
-        .with_input(Bytes::new())
-        .with_shared_address_generator(address_generator)
-        .with_chain_name(DEFAULT_CHAIN_NAME)
-        .with_block_time(Timestamp::now().into())
-        .with_state_hash(state_root_hash)
-        .with_block_height(1)
-        .with_parent_block_hash(BlockHash::new(Digest::hash(b"block1")))
-        .with_runtime_native_config(runtime_native_config)
-        .build()
-        .expect("should build");
-
-    let result = run_wasm_session(
-        &mut executor,
-        &global_state,
-        state_root_hash,
-        execute_request,
-    );
-
-    assert!(result.is_ok());
-}
-
-#[test]
-fn should_run_ee_966_regression_fail_when_growing_mem_past_max() {
-    let session_code = read_wasm(CONTRACT_EE_966_REGRESSION);
-
-    let mut executor = make_executor();
-    let (global_state, state_root_hash, _tempdir) = make_global_state_with_genesis();
-    let address_generator = make_address_generator();
-
-    let runtime_native_config = RuntimeNativeConfig::from_chainspec(&Chainspec::default());
-    let execute_request = ExecuteRequestBuilder::default()
-        .with_initiator(*DEFAULT_ACCOUNT_HASH)
-        .with_caller_key(Key::Account(*DEFAULT_ACCOUNT_HASH))
-        .with_gas_limit(DEFAULT_GAS_LIMIT)
-        .with_transferred_value(0)
-        .with_transaction_hash(TRANSACTION_HASH)
-        .with_target(ExecutionKind::SessionBytes(session_code))
-        .with_input(Bytes::new())
-        .with_shared_address_generator(address_generator)
-        .with_chain_name(DEFAULT_CHAIN_NAME)
-        .with_block_time(Timestamp::now().into())
-        .with_state_hash(state_root_hash)
-        .with_block_height(1)
-        .with_parent_block_hash(BlockHash::new(Digest::hash(b"block1")))
-        .with_runtime_native_config(runtime_native_config)
-        .build()
-        .expect("should build");
-
-    let result = run_wasm_session(
-        &mut executor,
-        &global_state,
-        state_root_hash,
-        execute_request,
-    );
-
-    println!("{:?}", result);
-
-    assert!(matches!(
-        result,
-        Ok(ExecuteWithProviderResult {
-            host_error: Some(CallError::CalleeReverted),
-            ..
-        })
-    ));
 }
